Create new projects, Infrastructure and Application Start by moving Member logic to Application Move Database logic to Infrastructure project Delete Services project
