+++ DisJockey.Application/Interfaces/IPlaylistRepository.cs
