+++ DisJockey.Application/Interfaces/IVideoDetailService.cs
