+++ DisJockey.Application/Members/Queries/AllMembers/AllMembersQuery.cs
@@ -1,0 +1,6 @@
+﻿using DisJockey.Shared.DTOs.Member;
+using MediatR;
+
+namespace DisJockey.Application.Members.Queries.AllMembers;
+
+public record AllMembersQuery : IRequest<IEnumerable<MemberListDto>>;
