+++ DisJockey.Infrastructure/SeedData.cs
@@ -1,0 +1,9 @@
+using DisJockey.Core;
+
+namespace DisJockey.Infrastructure;
+
+public class SeedData
+{
+    public ICollection<AppUser> Users { get; set; }
+    public ICollection<Track> Tracks { get; set; }
+}
