--- DisJockey.Services/DisJockey.Services.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Google.Apis.YouTube.v3" Version="1.52.0.2378" />
-        <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="5.0.8" />
-        <PackageReference Include="Microsoft.AspNetCore.WebUtilities" Version="2.2.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\DisJockey.Core\DisJockey.Core.csproj" />
-        <ProjectReference Include="..\DisJockey.Shared\DisJockey.Shared.csproj" />
-    </ItemGroup>
-
-</Project>
