+++ DisJockey.sln
@@ -9,8 +9,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DisJockey.Core", "DisJockey
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DisJockey.Shared", "DisJockey.Shared\DisJockey.Shared.csproj", "{F61F9FFE-B59D-45F1-BFB2-D878B047C0A3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DisJockey.Services", "DisJockey.Services\DisJockey.Services.csproj", "{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{69965E91-C2F3-4D6C-9251-2019D3BDAFBC}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
@@ -22,6 +20,14 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DisJockey.BotService", "DisJockey.BotService\DisJockey.BotService.csproj", "{3A0F2496-2629-46F5-A2E5-E92822017E8D}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DisJockey.MassTransit", "DisJockey.MassTransit\DisJockey.MassTransit.csproj", "{2EF66655-8FB3-4494-932F-8A4984B4ADC0}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{8D90E720-F214-4E3B-A948-C05BC1026A2B}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{2828C5AD-CC54-4E20-ADFF-1DA61FFAB616}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DisJockey.Application", "DisJockey.Application\DisJockey.Application.csproj", "{25A30481-6EC0-464A-A34D-4A67261FE0E8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DisJockey.Infrastructure", "DisJockey.Infrastructure\DisJockey.Infrastructure.csproj", "{16D380A1-AD31-45D8-933A-3D93433EACCC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -69,18 +75,6 @@ Global
 		{F61F9FFE-B59D-45F1-BFB2-D878B047C0A3}.Release|x64.Build.0 = Release|Any CPU
 		{F61F9FFE-B59D-45F1-BFB2-D878B047C0A3}.Release|x86.ActiveCfg = Release|Any CPU
 		{F61F9FFE-B59D-45F1-BFB2-D878B047C0A3}.Release|x86.Build.0 = Release|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Debug|x64.Build.0 = Debug|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Debug|x86.Build.0 = Debug|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Release|x64.ActiveCfg = Release|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Release|x64.Build.0 = Release|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Release|x86.ActiveCfg = Release|Any CPU
-		{3398DBE9-8B86-4269-A4A4-DD9BB60FA7B6}.Release|x86.Build.0 = Release|Any CPU
 		{EA706C75-32BE-43BE-97F5-E97555582290}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EA706C75-32BE-43BE-97F5-E97555582290}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EA706C75-32BE-43BE-97F5-E97555582290}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -117,9 +111,42 @@ Global
 		{2EF66655-8FB3-4494-932F-8A4984B4ADC0}.Release|x64.Build.0 = Release|Any CPU
 		{2EF66655-8FB3-4494-932F-8A4984B4ADC0}.Release|x86.ActiveCfg = Release|Any CPU
 		{2EF66655-8FB3-4494-932F-8A4984B4ADC0}.Release|x86.Build.0 = Release|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Debug|x64.Build.0 = Debug|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Debug|x86.Build.0 = Debug|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Release|x64.ActiveCfg = Release|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Release|x64.Build.0 = Release|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Release|x86.ActiveCfg = Release|Any CPU
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8}.Release|x86.Build.0 = Release|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Debug|x64.Build.0 = Debug|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Debug|x86.Build.0 = Debug|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Release|x64.ActiveCfg = Release|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Release|x64.Build.0 = Release|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Release|x86.ActiveCfg = Release|Any CPU
+		{16D380A1-AD31-45D8-933A-3D93433EACCC}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{0E7691AA-E6F1-41E2-86B4-5178B87C568D} = {8D90E720-F214-4E3B-A948-C05BC1026A2B}
+		{FC9DBC6E-9C36-4366-8511-83D3109F3B7E} = {8D90E720-F214-4E3B-A948-C05BC1026A2B}
+		{F61F9FFE-B59D-45F1-BFB2-D878B047C0A3} = {8D90E720-F214-4E3B-A948-C05BC1026A2B}
+		{3A0F2496-2629-46F5-A2E5-E92822017E8D} = {8D90E720-F214-4E3B-A948-C05BC1026A2B}
+		{2EF66655-8FB3-4494-932F-8A4984B4ADC0} = {8D90E720-F214-4E3B-A948-C05BC1026A2B}
+		{25A30481-6EC0-464A-A34D-4A67261FE0E8} = {8D90E720-F214-4E3B-A948-C05BC1026A2B}
+		{16D380A1-AD31-45D8-933A-3D93433EACCC} = {8D90E720-F214-4E3B-A948-C05BC1026A2B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1A465C47-4EF1-45C7-9E1A-C21BCCB31E5D}
